requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
requests==2.26.0
opencv-python==4.4.0.46
bunch==1.0.1
azureml-sdk
scikit-image==0.18.3
tensorflow
